python/requirements.txt
bluetooth-adapters
protobuf
grpcio
grpcio-tools
grpc-interceptor
bleak
cryptography
netifaces
zeroconf